move RulePostProcessor
