+++ src/main/java/org/delia/runner/TypeRunner.java
@@ -8,6 +8,7 @@ import org.delia.compiler.ast.TypeStatementExp;
 import org.delia.core.FactoryService;
 import org.delia.core.ServiceBase;
 import org.delia.error.DeliaError;
+import org.delia.rule.rules.RulePostProcessor;
 import org.delia.type.DType;
 import org.delia.type.DTypeRegistry;
 import org.delia.typebuilder.FutureDeclError;
